Add support for virtual offsets in the assembler
This makes it possible to place symbols outside of ROM space, for when a function is defined in ROM, but will be moved to RAM or some other memory later. The symbols will be generated in the ROM code as if the symbols and symbol references were already at that location.